% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cost_data.R
\name{get_cost_estimate}
\alias{get_cost_estimate}
\title{Get costs estimates}
\usage{
get_cost_estimate(x)
}
\arguments{
\item{x}{the imported database as returned by \link{import_data}}
}
\description{
This processes the 10 columns containing cost information for each row of the
original database.
}
