% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cost_data.R
\name{get_cost_smry}
\alias{get_cost_smry}
\title{Get costs summary}
\usage{
get_cost_smry(x)
}
\arguments{
\item{x}{the imported database as returned by \link{import_data}}
}
\description{
Obtain a text summary of cost estimates. This processes the 10 columns
containing cost information for each row of the original database.
}
