% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_data.R
\name{get_end_date}
\alias{get_end_date}
\title{Get ending date of outbreak}
\usage{
get_end_date(x)
}
\arguments{
\item{x}{the imported database as returned by \link{import_data}}
}
\description{
Get ending date of outbreak
}
