% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_data.R
\name{get_locations}
\alias{get_locations}
\title{Get country information}
\usage{
get_locations(x, concatenate = FALSE)
}
\arguments{
\item{x}{the imported database as returned by \link{import_data}}

\item{concatenate}{a \code{logical} indicating if several countries should be
collapsed into a single character string}
}
\description{
This can be one or more countries.
}
