% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_data.R
\name{get_outbreak_data}
\alias{get_outbreak_data}
\title{Get outbreak data}
\usage{
get_outbreak_data(x)
}
\arguments{
\item{x}{the imported database as returned by \link{import_data}}
}
\description{
Get outbreak data
}
