% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_data.R
\name{get_start_date}
\alias{get_start_date}
\title{Get starting date of outbreak}
\usage{
get_start_date(x)
}
\arguments{
\item{x}{the imported database as returned by \link{import_data}}
}
\description{
Get starting date of outbreak
}
