% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_numbers.R
\name{label_numbers}
\alias{label_numbers}
\title{Format numbers into labels K, M, etc.}
\usage{
label_numbers(x, show_1_as_0 = FALSE)
}
\arguments{
\item{x}{a \code{numeric} vector for which labels should be generated}

\item{show_1_as_0}{a \code{logical} indicating whether 1s should be shown as 0s;
this is only really useful when a log-transformation has been used and 1s
are in fact 0s.}
}
\description{
This function is used to make more human-readable labels for large numbers.
}
